Memory map: space for mapping nested page tables
We need some quick mechanism to access the nested page tables
for a given VM.  We do this by reserving another 512GiB region
of the host virtual address space to set up mappings _to the
nested page table structures_, but not the pages those point to.

The upshot of this is that we can use this region to access the
nested page tables in a manner analogous to how we access the
recursive page map, so that the paging structures in the EPT/NPT
are at known virtual addresses that correspond to their level in
the radix tree.  This allows us to quickly access those
structures without temporary mappings or a direct map.

What about the memory cost?  Consider that we are only mapping
the page table structures.  Those structures are already
allocated and assigned to the EPT/NPT anyway; we merely point to
them. Thus, using EPTs as an example, the PML1 entries in the
EPT are the target pages in our PML1; for each 4KiB page
allocated to the guest, this occupies 8 bytes in the EPT PML1.
But each EPT PML1 only occupies 8 bytes in our PML4, so memory
usage for these structures is reduced by a factor of ~2^28.
Paradoxically, levels higher in the EPT structure require more
space to map, but there are so many fewer of these that they are
insigificant.

Signed-off-by: Dan Cross <cross@gajendra.net>