xtask: Use the environment to find `llvm-objcopy`
This is a quality-of-life change only.

As part of the `xtask dist` process, we need to run
`objcopy` to rebrand the ELF executable as a 32-bit
bootable kernel image so that it can be understood
by a multiboot loader.  We usually use `llvm-objcopy`
from the `llvm-tools-preview` rustup component to do
this, but finding that can be a pain; previously,
I had been using an environment variable I expected
to simply be ambiently in the xtask process's
environment.  This change adds some logic to try and
find `llvm-objcopy` in the user's rustup root using
rustup variables.

Signed-off-by: Dan Cross <cross@gajendra.net>