Simplify libhypatia::runtime.
Now that the runtime generation macro only
generates a panic handler, generation for
segments and tasks is identical.  Remove
the two cases, and just call a single macro.

Also, make some minor aesthetic changes in
terms of placement in main.rs (at the end
of the file) and name of the associated
module (no underbar prefices; namespaces
obviate them).

Signed-off-by: Dan Cross <cross@gajendra.net>