pc: introduce a platform boundary
Right now, Theon is more or less specialized for use
on a platform that supplies ACPI, a multiboot loader,
etc.  This platform is, broadly speaking, derived
from the "PC", so we call it "pc".

However, we hope would like to be able to run on other
platforms, such as the Oxide platform, which does not
support ACPI or invoke a multiboot loader, etc.

So we start to introduce a platform abstraction, by
moving (most of) the PC-specific files into a pc/
subdirectory nested under `x86_64`.

Signed-off-by: Dan Cross <cross@gajendra.net>