theon: use `.space` for `bootstack`
Previously, `bootstack` (the initial stack for theon
when coming out of early boot) was allocated using
`.comm`, to place it into a COMMON section (a la
FORTRAN), which is known to go into the BSS.  However,
it would be lumped in with other COMMON objects, and
thus not necessarily aligned.  This isn't a huge deal,
but we can force alignment to a 4KiB boundary by
allocating it with, `.space` instead of `.comm`.

Signed-off-by: Dan Cross <cross@gajendra.net>