xtask: use `derive`-based command lines
Switch to using the `derive(Parser)` pattern in `xtask`.
This makes this code substantially more tidy and, I think,
easier to maintain and reason about.

Also add support for locking the `Cargo.lock` file on a
variety of commands, add a subcommand to expand macros,
add the ability to give number of CPUs and RAM size when
running qemu, and clean up creation of command lines.

Signed-off-by: Dan Cross <cross@gajendra.net>