rust: update rust, dependencies; fix errors
The ability to take a reference to a `static mut` is going to
become a hard error in Rust 2024.  This change attempts to get
out in front of that by modifying places where we do that to
make them safer.

The biggest change was in the bump allocator; the implementation
now is much closer to the one in the documentation for the
GlobalAlloc trait, in that it takes ownership of an (aligned)
array now, as opposed to wrapping a `static mut`.

Signed-off-by: Dan Cross <cross@gajendra.net>